обьединил в один солюшен, проверил все работает.

+++ src/MyRecommendationsService.Application/Interface/IVkApiServiceMyRecom.cs
@@ -4,6 +4,6 @@ namespace MyRecommendationsService.Application.Interface
 {
     public interface IVkApiServiceMyRecom
     {
-        Task<List<VkNet.Model.Attachments.Audio>> GetAudiosAsync(uint count, uint UserId);
+        Task<List<VkNet.Model.Attachments.Audio>> GetAudiosAsync(decimal count, decimal UserId);
     }
 }
